# cairn-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Cairn API
  description: >-
    Contract-first API of Cairn, a single-owner wealth tracker. It values a portfolio of accounts
    (envelopes such as a PEA or a savings account) and their holdings of instruments, priced from
    Yahoo Finance, CoinGecko, Societe Generale Sirius, or entered manually. Reading conventions:
    every amount is expressed in EUR; monetary amounts are rounded to 2 decimals on the wire;
    ratios are fractions, so 0.0082 means 0.82%; `share` values are fractions of a total and sum
    to 1 across a breakdown; a value that cannot be computed is absent from the response rather
    than sent as 0; dates are ISO-8601 (`2026-09-25`) and instants are UTC ISO-8601 timestamps.
  version: 0.1.0
...
security:
  - sessionCookie: []
paths:
...
  /instruments:
    get:
      tags: [Instrument]
      operationId: listInstruments
      summary: List every instrument
      description: Every instrument known to Cairn, whether or not any account currently holds it.
      responses:
        '200':
          description: Every instrument
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstrumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Instrument]
      operationId: createInstrument
      summary: Create an instrument
      description: The ISIN and source reference, when given, must each be unique across every instrument.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstrumentRequest'
            example:
              name: Sample S&P 500 ETF
              isin: FR0011550185
              currency: EUR
              assetClass: ETF
              priceSource: YAHOO
              sourceRef: SXR8.DE
      responses:
        '201':
          description: Instrument created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Another instrument already uses this ISIN or source reference
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: >-
            Invalid instrument: `sourceRef` is required unless `priceSource` is `MANUAL`, or
            `description` is over 280 characters
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /instruments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The instrument's id.
        schema:
          type: string
          format: uuid
    get:
      tags: [Instrument]
      operationId: getInstrument
      summary: Get an instrument's detail
      description: The full record, including the fields not shown in the list view, plus the number of holdings that would be deleted along with it.
      responses:
        '200':
          description: Instrument detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Unknown instrument
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags: [Instrument]
      operationId: updateInstrument
      summary: Replace an instrument's editable fields
      description: The currency is not editable; it is kept from the existing instrument.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstrumentRequest'
            example:
              name: Sample S&P 500 ETF
              assetClass: ETF
              priceSource: YAHOO
              isin: FR0011550185
              sourceRef: SXR8.DE
      responses:
        '200':
          description: Instrument updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Unknown instrument
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Another instrument already uses this ISIN or source reference
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: >-
            Invalid instrument: `sourceRef` is required unless `priceSource` is `MANUAL`, or
            `description` is over 280 characters. Attempting to change the currency is silently
            ignored rather than rejected.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      tags: [Instrument]
      operationId: deleteInstrument
      summary: Delete an instrument and its holdings
      description: Deletes the instrument and every holding of it, in one transaction.
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        '204':
          description: Instrument deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Unknown instrument
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /instruments/search:
    get:
      tags: [Instrument]
      operationId: searchInstruments
      summary: Search one price source for instruments
      description: >-
        One source per call, so each source gets its own loading, empty and error state. The web picks
        the sources: by default YAHOO and COINGECKO, plus AMUNDI when the query is a full ISIN; an ISIN
        starting with QS queries AMUNDI only; an ISIN never queries COINGECKO. AMUNDI answers a full
        12-character ISIN only and returns an empty list otherwise. Each candidate is probed for a live
        price. A candidate already tracked (same source and reference) carries `trackedInstrumentId`.
        An empty list means the source knows nothing; a source that is down answers 502.
      parameters:
        - name: source
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/SearchableSource'
        - name: query
          in: query
          required: true
          description: An ISIN, a ticker or a name
          schema:
            type: string
            minLength: 2
            maxLength: 64
            example: FR0011550185
      responses:
        '200':
          description: Candidates of this source, each probed for a live price, possibly none
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstrumentCandidateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          description: The source did not answer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: >-
        Session opened by a passkey (WebAuthn) sign-in, lasting 30 days, sliding. The ceremony
        itself is served by Spring Security, not by this contract: `POST /webauthn/register/options`
        then `POST /webauthn/register` for registration, `POST /webauthn/authenticate/options`
        then `POST /login/webauthn` for sign-in, and `POST /logout` to end the session.
    csrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: >-
        Required on every POST, PUT, PATCH and DELETE. The value is the raw token from the
        `XSRF-TOKEN` cookie the server sets; a write submitted without it is refused with 403.
  responses:
    Unauthorized:
      description: No valid session
    Forbidden:
      description: Missing or wrong CSRF token on a write
    BadRequest:
      description: >-
        The body could not be read, a required field is missing, or a parameter has the wrong
        type. Rendered in Spring Boot's default error format, not as a problem detail.
...
  schemas:
    ProblemDetail:
      description: >-
        RFC 9457 problem detail, served as `application/problem+json` for every business, conflict
        and not-found error.
      type: object
      properties:
        type:
          type: string
          description: A URI reference identifying the problem type. `about:blank` when none is more specific.
          example: about:blank
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: Unprocessable Entity
        status:
          type: integer
          description: The HTTP status code, repeated from the response.
          example: 422
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: account 6f1c2d9e-4b7a-4c1e-9a53-0d2f8a1b7e44 already holds instrument 3b9d7c10-52e8-4f6a-8c2b-91a4d0e5f7a2
        instance:
          type: string
          description: A URI reference identifying this specific occurrence of the problem.
          example: /holdings
...
    AssetClass:
      type: string
      description: >-
        The kind of instrument. `EQUITY` is a single stock, `ETF` an exchange-traded fund, `FUND`
        a non-traded fund, `CRYPTO` a cryptocurrency, and `CASH` a currency balance priced at par.
      enum: [EQUITY, ETF, FUND, CRYPTO, CASH]
    PriceSource:
      type: string
      description: >-
        Where an instrument's price comes from. `YAHOO` is Yahoo Finance, `COINGECKO` is CoinGecko,
        `SG_SIRIUS` is Societe Generale's structured-product pricing feed, `AMUNDI` is Amundi's fund
        data repository (Amundi, Societe Generale Gestion and the employee savings and retirement funds
        they manage), and `MANUAL` means the price is entered by hand rather than fetched.
      enum: [YAHOO, COINGECKO, SG_SIRIUS, AMUNDI, MANUAL]
...
    InstrumentResponse:
      description: >-
        A tradable or holdable instrument as it appears in a list, priced from a market data
        source or entered manually. Returned by `GET /instruments` and the instrument write
        endpoints.
      type: object
      required: [id, name, currency, assetClass, priceSource]
      properties:
        id:
          type: string
          format: uuid
          description: The instrument's id.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          description: The instrument's display name.
          example: Sample S&P 500 ETF
        isin:
          type: string
          nullable: true
          description: The instrument's ISIN. Absent when the instrument has none.
          example: FR0011550185
        symbol:
          type: string
          nullable: true
          description: The instrument's ticker as people read it. Absent when the instrument has none.
          example: AAPL
        currency:
          type: string
          description: The instrument's native pricing currency, as an ISO 4217 code.
          example: EUR
        assetClass:
          description: The kind of instrument.
          allOf:
            - $ref: '#/components/schemas/AssetClass'
        priceSource:
          description: Where the instrument's price comes from.
          allOf:
            - $ref: '#/components/schemas/PriceSource'
        sourceRef:
          type: string
          nullable: true
          description: The ticker, contract identifier or ISIN the price source is queried with. Absent for MANUAL.
          example: SXR8.DE
    InstrumentDetailResponse:
      description: >-
        The full detail of one instrument, including fields not shown in the list view and the
        number of holdings that would be deleted with it. Returned by `GET /instruments/{id}` and
        the instrument write endpoints.
      type: object
      required: [id, name, currency, assetClass, priceSource, holdingCount]
      properties:
        id:
          type: string
          format: uuid
          description: The instrument's id.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          description: The instrument's display name.
          example: Sample S&P 500 ETF
        isin:
          type: string
          nullable: true
          description: The instrument's ISIN. Absent when the instrument has none.
          example: FR0011550185
        symbol:
          type: string
          nullable: true
          description: The instrument's ticker as people read it. Absent when the instrument has none.
          example: AAPL
        currency:
          type: string
          description: The instrument's native pricing currency, as an ISO 4217 code. Not editable.
          example: EUR
        assetClass:
          description: The kind of instrument.
          allOf:
            - $ref: '#/components/schemas/AssetClass'
        priceSource:
          description: Where the instrument's price comes from.
          allOf:
            - $ref: '#/components/schemas/PriceSource'
        sourceRef:
          type: string
          nullable: true
          description: The ticker, contract identifier or ISIN the price source is queried with. Absent for MANUAL.
          example: SXR8.DE
        description:
          type: string
          nullable: true
          description: Free text, up to 280 characters.
          example: Core equity ETF tracking the S&P 500.
        externalUrl:
          type: string
          nullable: true
          description: Absent for a manually priced instrument and for an Amundi fund, since Amundi has no public page for every share. Never a broken link.
          example: https://finance.yahoo.com/quote/SXR8.DE
        holdingCount:
          type: integer
          description: The number of holdings that will be deleted with the instrument.
          example: 2
    CreateInstrumentRequest:
      description: The body of `POST /instruments`.
      type: object
      required: [name, currency, assetClass, priceSource]
      properties:
        name:
          type: string
          description: The instrument's display name.
          example: Sample S&P 500 ETF
        isin:
          type: string
          nullable: true
          description: The instrument's ISIN. Must be unique across every instrument when given.
          example: FR0011550185
        symbol:
          type: string
          maxLength: 32
          nullable: true
          description: The instrument's ticker as people read it, for example the candidate's `symbol`. Free text up to 32 characters.
          example: AAPL
        currency:
          type: string
          description: The instrument's native pricing currency, as an ISO 4217 code.
          example: EUR
        assetClass:
          description: The kind of instrument.
          allOf:
            - $ref: '#/components/schemas/AssetClass'
        priceSource:
          description: Where the instrument's price comes from.
          allOf:
            - $ref: '#/components/schemas/PriceSource'
        sourceRef:
          type: string
          nullable: true
          description: Required unless priceSource is MANUAL.
          example: SXR8.DE
        description:
          type: string
          nullable: true
          description: Free text, up to 280 characters.
          example: Core equity ETF tracking the S&P 500.
    UpdateInstrumentRequest:
      description: The body of `PUT /instruments/{id}`. The currency cannot be changed this way.
      type: object
      required: [name, assetClass, priceSource]
      properties:
        name:
          type: string
          description: The instrument's display name.
          example: Sample S&P 500 ETF
        assetClass:
          description: The kind of instrument.
          allOf:
            - $ref: '#/components/schemas/AssetClass'
        priceSource:
          description: Where the instrument's price comes from.
          allOf:
            - $ref: '#/components/schemas/PriceSource'
        isin:
          type: string
          nullable: true
          description: The instrument's ISIN. Must be unique across every instrument when given.
          example: FR0011550185
        symbol:
          type: string
          maxLength: 32
          nullable: true
          description: The instrument's ticker as people read it. Replaced as sent, so omitting it clears the stored symbol. Up to 32 characters.
          example: AAPL
        sourceRef:
          type: string
          nullable: true
          description: Required unless priceSource is MANUAL.
          example: SXR8.DE
        description:
          type: string
          nullable: true
          description: Free text, up to 280 characters.
          example: Core equity ETF tracking the S&P 500.
    SearchableSource:
      type: string
      enum: [YAHOO, COINGECKO, AMUNDI]
      description: The price sources `GET /instruments/search` can query.
    InstrumentCandidateResponse:
      description: >-
        One price source's answer while resolving a query, with a live price fetched as a sanity
        check. Returned by `GET /instruments/search`.
      type: object
      required: [name, source, sourceRef, assetClass]
      properties:
        name:
          type: string
          description: The instrument's name as reported by this source.
          example: iShares Core S&P 500 UCITS ETF
        source:
          description: The price source this candidate was resolved from.
          allOf:
            - $ref: '#/components/schemas/PriceSource'
        sourceRef:
          type: string
          description: The ticker, contract identifier or ISIN this candidate would be resolved with.
          example: SXR8.DE
        assetClass:
          description: The kind of instrument this source reports it as.
          allOf:
            - $ref: '#/components/schemas/AssetClass'
        isin:
          type: string
          nullable: true
          description: >-
            The instrument's ISIN when the source can tell it, else null. Yahoo Finance does not
            return it, so it is set only when the query was itself an ISIN, which the source then matched.
            Amundi candidates always carry it.
          example: US0378331005
        symbol:
          type: string
          nullable: true
          description: >-
            The ticker as the source gives it, exchange suffix included ("AI.PA"). Null when the
            source gives none.
          example: AI.PA
        exchange:
          type: string
          nullable: true
          description: The exchange this candidate is listed on, as the source names it. Null when the source does not say.
          example: Paris
        probePrice:
          type: number
          nullable: true
          description: >-
            The live price fetched from this source while resolving, as a sanity check. Null when
            the probe failed or timed out.
          example: 28.10
        currency:
          type: string
          nullable: true
          description: >-
            The currency the probe quote is expressed in (ISO 4217). Null when the probe failed or
            timed out, in which case the candidate stays selectable. Candidates are sorted EUR first,
            then unknown, then other currencies, in the source's order within each group.
          example: EUR
        probeAsOf:
          type: string
          format: date
          nullable: true
          description: >-
            The date of the probe price (Amundi NAV date, CoinGecko last update, quote date for
            Yahoo Finance). Null when the probe failed.
          example: 2026-09-24
        trackedInstrumentId:
          type: string
          format: uuid
          nullable: true
          description: >-
            The instrument already tracked with this source and reference, else null; the web lists
            such a candidate under "Déjà suivi".
...